% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64}
\alias{base64}
\alias{base64_decode}
\alias{base64_encode}
\title{Encode/decode base64}
\usage{
base64_decode(input)

base64_encode(input)
}
\arguments{
\item{input}{string or raw vector to be encoded/decoded}
}
\description{
Simple in-memory base64 encoder and decoder. Interchangable with
implementations from \code{base64enc} or \code{openssl} packages.
}

