% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.pages.R
\name{rbind.pages}
\alias{rbind.pages}
\title{Deprecated}
\usage{
\method{rbind}{pages}(...)
}
\arguments{
\item{...}{see \link{rbind_pages}}
}
\description{
This function has been renamed to \link{rbind_pages} due to changes in R.
}
