% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asJSON.vctrs.R
\name{vec_proxy_json}
\alias{vec_proxy_json}
\title{JSON conversion proxy}
\usage{
vec_proxy_json(x)
}
\arguments{
\item{x}{object instance of class \code{vctrs_vctr} to be converted to JSON}
}
\value{
an atomic vector or a list of atomic vectors.
}
\description{
vec_proxy_json method returns proxy objects, i.e. an atomic vector or a list of atomic vectors,
able to be converted to JSON using the jsonlite package. This is a generic S3 method, that has
to be implemented for subclasses requesting particualar JSON conversion. A default implementation
is provided for \code{vctrs_vctr} class.
}
