% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.json.R
\name{write.json}
\alias{write.json}
\title{Wrapper around toJSON.}
\usage{
write.json(x, file = "", ...)
}
\arguments{
\item{x}{the object to be encoded}

\item{file}{a JSON file}

\item{...}{arguments passed on to \code{\link{toJSON}}}
}
\description{
Wrapper around toJSON.
}
\examples{
\dontrun{

write.json(x = data, file="data.json")
}
}

