% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64}
\alias{base64}
\alias{base64_dec}
\alias{base64_enc}
\title{Encode/decode base64}
\usage{
base64_dec(input)

base64_enc(input)
}
\arguments{
\item{input}{string or raw vector to be encoded/decoded}
}
\description{
Simple in-memory base64 encoder and decoder. Used internally for converting
raw vectors to text. Interchangeable with encoder from \code{base64enc} or
\code{openssl} package.
}
\examples{
str <- base64_enc(serialize(iris, NULL))
out <- unserialize(base64_dec(str))
stopifnot(identical(out, iris))
}
