% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prettify, minify}
\alias{minify}
\alias{prettify}
\alias{prettify, minify}
\title{Prettify or minify a JSON string}
\usage{
prettify(txt, indent = 4)

minify(txt)
}
\arguments{
\item{txt}{JSON string}

\item{indent}{number of spaces to indent}
}
\description{
Prettify adds indentation to a JSON string; minify removes all indentation/whitespace.
}
\examples{
myjson <- toJSON(cars)
cat(myjson)
prettify(myjson)
minify(myjson)
}

