% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prettify, minify}
\alias{minify}
\alias{prettify}
\alias{prettify, minify}
\title{Prettify or minify a JSON string}
\usage{
prettify(txt)

minify(txt)
}
\arguments{
\item{txt}{JSON string}
}
\description{
Prettify adds indentation to a JSON string; minify removes all indentation/whitespace.
}
\examples{
myjson <- toJSON(cars)
cat(myjson)
prettify(myjson)
minify(myjson)
}

