% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify, minify}
\alias{prettify, minify}
\alias{prettify}
\alias{minify}
\title{Prettify or minify a JSON string}
\usage{
prettify(txt, indent = 4)

minify(txt)
}
\arguments{
\item{txt}{JSON string}

\item{indent}{number of spaces to indent. Use a negative number for tabs instead of spaces.}
}
\description{
Prettify adds indentation to a JSON string; minify removes all indentation/whitespace.
}
\examples{
myjson <- toJSON(cars)
cat(myjson)
prettify(myjson)
minify(myjson)
}
