% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toMongo}
\alias{toMongo}
\title{Export to MongoDB}
\usage{
toMongo(x, array = FALSE, POSIXt = "mongo", raw = "mongo", ...)
}
\arguments{
\item{x}{a dataframe}

\item{array}{if output should be an array. See details.}

\item{POSIXt}{passed on to \code{\link{toJSON}}}

\item{raw}{passed on to \code{\link{toJSON}}}

\item{...}{other arguments for \code{\link{toJSON}}}
}
\description{
This function will be deprecated. Please use stream_out instead.
}
\details{
This function will be deprecated. Please use stream_out instead.
}

