% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toMongo}
\alias{toMongo}
\title{Export to MongoDB}
\usage{
toMongo(x, array = FALSE, POSIXt = "mongo", raw = "mongo", ...)
}
\arguments{
\item{x}{a dataframe}

\item{array}{if output should be an array. See details.}

\item{POSIXt}{passed on to \code{\link{toJSON}}}

\item{raw}{passed on to \code{\link{toJSON}}}

\item{...}{other arguments for \code{\link{toJSON}}}
}
\description{
Exports a dataframe to a format that can be imported with the
\code{mongoimport} command line tool included with MongoDB.
}
\details{
By default \code{mongoimport} expects a datafile in which every line is a record
in the collection. Hence the complete output is not valid JSON itself.
Alternatively, if \code{array=TRUE} the output will be wrapped a JSON array.
When using the latter, we need to pass \code{--jsonArray} to \code{mongoimport}.
}

