% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate JSON}
\usage{
validate(txt)
}
\arguments{
\item{txt}{JSON string}
}
\description{
Test if a string contains valid JSON. Characters vectors will be collapsed into a single string.
}
\examples{
#Output from toJSON and serializeJSON should pass validation
myjson <- toJSON(mtcars)
validate(myjson) #TRUE

#Something bad happened
truncated <- substring(myjson, 1, 100)
validate(truncated) #FALSE
}

